[#33][refact] split game - game related files moved to the same folder
