+++ front/public/game/components/game-renderer-2d.ce.js
